+++ src/Modules/OrchardCore.Commerce/Services/StripePaymentService.cs
@@ -31,9 +31,7 @@ public class StripePaymentService : IStripePaymentService
 {
     private readonly IShoppingCartHelpers _shoppingCartHelpers;
     private readonly IShoppingCartPersistence _shoppingCartPersistence;
-    private readonly IPriceService _priceService;
     private readonly PaymentIntentService _paymentIntentService;
-    private readonly IPriceSelectionStrategy _priceSelectionStrategy;
     private readonly IContentManager _contentManager;
     private readonly IStringLocalizer T;
     private readonly ISession _session;
@@ -43,14 +41,13 @@ public class StripePaymentService : IStripePaymentService
     private readonly IContentItemDisplayManager _contentItemDisplayManager;
     private readonly IProductInventoryService _productInventoryService;
     private readonly IEnumerable<IWorkflowManager> _workflowManagers;
+    private readonly IPaymentService _paymentService;
 
     // We need to use that many this cannot be avoided.
 #pragma warning disable S107 // Methods should not have too many parameters
     public StripePaymentService(
         IShoppingCartHelpers shoppingCartHelpers,
         IShoppingCartPersistence shoppingCartPersistence,
-        IPriceService priceService,
-        IPriceSelectionStrategy priceSelectionStrategy,
         IContentManager contentManager,
         ISiteService siteService,
         IDataProtectionProvider dataProtectionProvider,
@@ -60,18 +57,18 @@ public class StripePaymentService : IStripePaymentService
         IPaymentIntentPersistence paymentIntentPersistence,
         IContentItemDisplayManager contentItemDisplayManager,
         IProductInventoryService productInventoryService,
-        IEnumerable<IWorkflowManager> workflowManagers)
+        IEnumerable<IWorkflowManager> workflowManagers,
+        IPaymentService paymentService)
 #pragma warning restore S107 // Methods should not have too many parameters
     {
         _paymentIntentService = new PaymentIntentService();
         _shoppingCartHelpers = shoppingCartHelpers;
         _shoppingCartPersistence = shoppingCartPersistence;
-        _priceService = priceService;
-        _priceSelectionStrategy = priceSelectionStrategy;
         _contentManager = contentManager;
         _session = session;
         _paymentIntentPersistence = paymentIntentPersistence;
         _productInventoryService = productInventoryService;
+        _paymentService = paymentService;
         T = stringLocalizer;
         _contentItemDisplayManager = contentItemDisplayManager;
         _workflowManagers = workflowManagers;
@@ -235,7 +232,7 @@ public class StripePaymentService : IStripePaymentService
 
         order.DisplayText = T["Order {0}", guidId];
 
-        var lineItems = await CreateOrderLineItemsAsync(currentShoppingCart);
+        var lineItems = await _paymentService.CreateOrderLineItemsAsync(currentShoppingCart);
 
         var cartViewModel = await _shoppingCartHelpers.CreateShoppingCartViewModelAsync(
             shoppingCartId: null,
@@ -284,76 +281,6 @@ public class StripePaymentService : IStripePaymentService
         }
 
         return order;
-    }
-
-    public async Task<ContentItem> CreateNoPaymentOrderFromShoppingCartAsync(IUpdateModelAccessor updateModelAccessor)
-    {
-        var currentShoppingCart = await _shoppingCartPersistence.RetrieveAsync();
-
-        var order = await _contentManager.NewAsync(Constants.ContentTypes.Order);
-        if (await UpdateOrderWithDriversAsync(order, updateModelAccessor))
-        {
-            return null;
-        }
-
-        var guid = Guid.NewGuid().ToString();
-        order.DisplayText = T["Order {0}", guid];
-
-        var lineItems = await CreateOrderLineItemsAsync(currentShoppingCart);
-
-        var cartViewModel = await _shoppingCartHelpers.CreateShoppingCartViewModelAsync(
-            shoppingCartId: null,
-            order.As<OrderPart>().ShippingAddress.Address,
-            order.As<OrderPart>().BillingAddress.Address);
-
-        order.Alter<OrderPart>(orderPart =>
-        {
-            // Shopping cart
-            orderPart.LineItems.Clear();
-            orderPart.LineItems.AddRange(lineItems);
-
-            orderPart.OrderId.Text = guid;
-            orderPart.Status = new TextField { ContentItem = order, Text = OrderStatuses.Pending.HtmlClassify() };
-
-            // Store the current applicable discount info so they will be available in the future.
-            orderPart.AdditionalData.SetDiscountsByProduct(cartViewModel
-                .Lines
-                .Where(line => line.AdditionalData.GetDiscounts().Any())
-                .ToDictionary(
-                    line => line.ProductSku,
-                    line => line.AdditionalData.GetDiscounts()));
-        });
-
-        await _contentManager.CreateAsync(order);
-
-        return order;
-    }
-
-    public async Task<IEnumerable<OrderLineItem>> CreateOrderLineItemsAsync(ShoppingCart shoppingCart)
-    {
-        var lineItems = new List<OrderLineItem>();
-
-        // This needs to be done separately because it's async: "Avoid using async lambda when delegate type returns
-        // void."
-        foreach (var item in shoppingCart.Items)
-        {
-            var trimmedSku = item.ProductSku.Split('-').First();
-
-            var contentItemId = (await _session
-                    .QueryIndex<ProductPartIndex>(productPartIndex => productPartIndex.Sku == trimmedSku)
-                    .ListAsync())
-                .Select(index => index.ContentItemId)
-                .First();
-
-            var contentItemVersion = (await _contentManager.GetAsync(contentItemId)).ContentItemVersionId;
-
-            lineItems.Add(await item.CreateOrderLineFromShoppingCartItemAsync(
-                _priceSelectionStrategy,
-                _priceService,
-                contentItemVersion));
-        }
-
-        return lineItems;
     }
 
     private async Task<bool> UpdateOrderWithDriversAsync(ContentItem order, IUpdateModelAccessor updateModelAccessor)
